Arduino Code for ML implementation using ESP32 Camera Module
To integrate ML in this,  we used an ESP32 camera modue using another arduino. After uploading the code the shorted connection between pin0 and gnd is removed.
The esp32 camera module is reset to get the web link in the serial monitor . This link is accessed in the ML code , which accesses this web for fetching continuous images 
and create a videostream. 
Can check the camera working by opening the web link .